Adds a fix to scale the Frequency dialog.
The dialog layout uses a sizer to scale the content automatically.  The layout wasn't very good using sizers alone as width of the dialog was being set too narrow for the title string.  The width is set manually to either the [title string width + 35%] or the [content string width +10%], whichever's the greater.